Architecture: remove viewModel and factory, moved keyple logic in CardReaderActivity
