--- validator-app/src/main/java/org/eclipse/keyple/famoco/validator/di/ViewModelFactory.java
@@ -1,44 +1,0 @@
-package org.eclipse.keyple.famoco.validator.di;
-
-import androidx.lifecycle.ViewModel;
-import androidx.lifecycle.ViewModelProvider;
-
-import org.eclipse.keyple.famoco.validator.di.scopes.AppScoped;
-
-import java.util.Map;
-
-import javax.inject.Inject;
-import javax.inject.Provider;
-
-//Wee keep this in java because of: https://github.com/google/dagger/issues/1478
-@AppScoped
-public class ViewModelFactory implements ViewModelProvider.Factory {
-    private final Map<Class<? extends ViewModel>, Provider<ViewModel>> creators;
-
-    @Inject
-    public ViewModelFactory(Map<Class<? extends ViewModel>, Provider<ViewModel>> creators) {
-        this.creators = creators;
-    }
-
-    @SuppressWarnings("unchecked")
-    @Override
-    public <T extends ViewModel> T create(Class<T> modelClass) {
-        Provider<? extends ViewModel> creator = creators.get(modelClass);
-        if (creator == null) {
-            for (Map.Entry<Class<? extends ViewModel>, Provider<ViewModel>> entry : creators.entrySet()) {
-                if (modelClass.isAssignableFrom(entry.getKey())) {
-                    creator = entry.getValue();
-                    break;
-                }
-            }
-        }
-        if (creator == null) {
-            throw new IllegalArgumentException("unknown model class " + modelClass);
-        }
-        try {
-            return (T) creator.get();
-        } catch (Exception e) {
-            throw new RuntimeException(e);
-        }
-    }
-}
