+++ validator-app/src/main/kotlin/org/eclipse/keyple/famoco/validator/di/AppComponent.kt
@@ -19,7 +19,7 @@ import org.eclipse.keyple.famoco.validator.Application
 import org.eclipse.keyple.famoco.validator.di.scopes.AppScoped
 
 @AppScoped
-@Component(modules = [ViewModelModule::class, AppModule::class, UIModule::class, SchedulerModule::class, AndroidSupportInjectionModule::class])
+@Component(modules = [AppModule::class, UIModule::class, SchedulerModule::class, AndroidSupportInjectionModule::class])
 interface AppComponent : AndroidInjector<Application?> {
     @Component.Builder
     interface Builder {
