--- validator-app/src/main/kotlin/org/eclipse/keyple/famoco/validator/di/ViewModelModule.kt
@@ -1,32 +1,0 @@
-/********************************************************************************
- * Copyright (c) 2020 Calypso Networks Association https://www.calypsonet-asso.org/
- *
- * See the NOTICE file(s) distributed with this work for additional information regarding copyright
- * ownership.
- *
- * This program and the accompanying materials are made available under the terms of the Eclipse
- * Public License 2.0 which is available at http://www.eclipse.org/legal/epl-2.0
- *
- * SPDX-License-Identifier: EPL-2.0
- ********************************************************************************/
-package org.eclipse.keyple.famoco.validator.di
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import dagger.Binds
-import dagger.Module
-import dagger.multibindings.IntoMap
-import org.eclipse.keyple.famoco.validator.di.scopes.AppScoped
-import org.eclipse.keyple.famoco.validator.viewModels.CardReaderViewModel
-
-@Module
-abstract class ViewModelModule {
-    @Binds
-    @IntoMap
-    @ViewModelKey(CardReaderViewModel::class)
-    abstract fun bindCardReaderViewModel(cardReaderViewModel: CardReaderViewModel?): ViewModel?
-
-    @Binds
-    @AppScoped
-    abstract fun bindViewModelFactory(factory: ViewModelFactory?): ViewModelProvider.Factory?
-}
