+++ validator-app/src/noFamoco/kotlin/org/eclipse/keyple/famoco/validator/ui/activities/CardReaderActivity.kt
@@ -26,52 +26,32 @@ package org.eclipse.keyple.famoco.validator.ui.activities
 import android.content.Intent
 import android.os.Bundle
 import android.view.View
-import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModelProvider
 import com.airbnb.lottie.LottieDrawable
 import dagger.android.support.DaggerAppCompatActivity
 import kotlinx.android.synthetic.main.activity_card_reader.*
-import org.eclipse.keyple.core.seproxy.exception.KeyplePluginInstantiationException
 import org.eclipse.keyple.famoco.validator.R
 import org.eclipse.keyple.famoco.validator.data.model.CardReaderResponse
 import org.eclipse.keyple.famoco.validator.data.model.Status
 import org.eclipse.keyple.famoco.validator.di.scopes.ActivityScoped
-import org.eclipse.keyple.famoco.validator.ui.BaseView
 import org.eclipse.keyple.famoco.validator.util.KeypleSettings
-import org.eclipse.keyple.famoco.validator.viewModels.CardReaderViewModel
-import timber.log.Timber
 import java.util.*
-import javax.inject.Inject
 
 @ActivityScoped
-class CardReaderActivity @Inject constructor() : DaggerAppCompatActivity(), BaseView {
-    @Inject
-    lateinit var viewModelFactory: ViewModelProvider.Factory
-    lateinit var cardReaderViewModel: CardReaderViewModel
+class CardReaderActivity: DaggerAppCompatActivity() {
 
     private val timer = Timer()
 
     override fun onCreate(savedInstanceState: Bundle?){
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_card_reader)
-        cardReaderViewModel = ViewModelProvider(this, viewModelFactory).get(CardReaderViewModel::class.java)
         // Inflate the layout for this fragment
         animation.setAnimation("card_scan.json")
         animation.playAnimation()
-
-        try {
-            cardReaderViewModel.initCardReader()
-        } catch (e: KeyplePluginInstantiationException) {
-            Timber.e(e)
-        }
     }
 
     override fun onResume() {
         super.onResume()
-        bindViewModel()
         animation.playAnimation()
-        cardReaderViewModel.startNfcDetection(this)
-
         if(KeypleSettings.batteryPowered) {
             timer.schedule(object : TimerTask() {
                 override fun run() {
@@ -79,22 +59,37 @@ class CardReaderActivity @Inject constructor() : DaggerAppCompatActivity(), Base
                 }
             }, RETURN_DELAY_MS.toLong())
         }
+
+        // STUB Card event
+        val timer = Timer()
+        timer.schedule(object : TimerTask() {
+            override fun run() {
+                /** Change this value to see other status screens **/
+                val status: Status = Status.INVALID_CARD
+                when (status) {
+                    Status.TICKETS_FOUND -> changeDisplay(
+                        CardReaderResponse(
+                        Status.TICKETS_FOUND,
+                        7,
+                        "Season Pass",
+                        ""
+                    ))
+                    Status.LOADING, Status.ERROR, Status.SUCCESS, Status.INVALID_CARD, Status.EMPTY_CARD -> changeDisplay(
+                        CardReaderResponse(
+                        status,
+                        0,
+                        "",
+                        ""
+                    ))
+                }
+            }
+        }, EVENT_DELAY_MS.toLong())
     }
 
     override fun onPause() {
         super.onPause()
-        unbindViewModel()
         animation.cancelAnimation()
-        cardReaderViewModel.stopNfcDetection(this)
         timer.cancel()
-    }
-
-    override fun bindViewModel() {
-        cardReaderViewModel.response.observe(this, Observer { cardReaderResponse: CardReaderResponse? -> changeDisplay(cardReaderResponse) })
-    }
-
-    override fun unbindViewModel() {
-        cardReaderViewModel.response.removeObservers(this)
     }
 
     private fun changeDisplay(cardReaderResponse: CardReaderResponse?) {
@@ -106,7 +101,9 @@ class CardReaderActivity @Inject constructor() : DaggerAppCompatActivity(), Base
                 animation.playAnimation()
                 animation.repeatCount = LottieDrawable.INFINITE
             } else {
-                animation.cancelAnimation()
+                runOnUiThread{
+                    animation.cancelAnimation()
+                }
                 val intent = Intent(this, CardSummaryActivity::class.java)
                 intent.putExtra(CardSummaryActivity.STATUS_KEY, cardReaderResponse.status.toString())
                 intent.putExtra(CardSummaryActivity.TICKETS_KEY, cardReaderResponse.ticketsNumber)
@@ -121,5 +118,6 @@ class CardReaderActivity @Inject constructor() : DaggerAppCompatActivity(), Base
 
     companion object {
         private const val RETURN_DELAY_MS = 30000
+        private const val EVENT_DELAY_MS = 3000
     }
 }
